feat: :zap: F64 conversion

+++ assembly/UniswapV3LiquidityStrategy.ts
@@ -1,11 +1,11 @@
 import { BaseStrategy } from '@steerprotocol/base-strategy/assembly';
 
 export class UniswapV3LiquidityStrategy extends BaseStrategy {
-  private LIQUIDITY_PRECISION: f32 = 2 ** 16 - 2; 
-  private startTick: f32;
-  public binWidth: f32 = 600;
+  private LIQUIDITY_PRECISION: i64 = 2 ** 16 - 2; 
+  private startTick: i32;
+  public binWidth: i32 = 600;
 
-  constructor(binWidth: f32) {
+  constructor(binWidth: i32) {
     super();
     this.binWidth = binWidth;
   }
